+++ addons/l10n_in/models/__init__.py
@@ -1,7 +1,8 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 from . import template_in
-from . import account
 from . import account_invoice
+from . import account_move_line
+from . import account_tax
 from . import company
 from . import product_template
 from . import port_code
