+++ addons/l10n_in/models/account_move_line.py
@@ -1,0 +1,15 @@
+from odoo import api, fields, models
+
+
+class AccountMoveLine(models.Model):
+    _inherit = "account.move.line"
+
+    l10n_in_hsn_code = fields.Char(string="HSN/SAC Code", compute="_compute_l10n_in_hsn_code", store=True, readonly=False, copy=False)
+
+    @api.depends('product_id')
+    def _compute_l10n_in_hsn_code(self):
+        indian_lines = self.filtered(lambda line: line.company_id.account_fiscal_country_id.code == 'IN')
+        (self - indian_lines).l10n_in_hsn_code = False
+        for line in indian_lines:
+            if line.product_id:
+                line.l10n_in_hsn_code = line.product_id.l10n_in_hsn_code
